Fix comment ajax for Service, Baptism, Funeral, Wedding models
