+++ app/Http/Controllers/DownloadController.php
@@ -24,9 +24,11 @@ class DownloadController extends Controller
 
     public function attachment(Request $request, Attachment $attachment, $prettyName = '') {
         if (get_class($attachment->attachable) == Service::class) {
+            if (!Auth::user()->can('update', $attachment->attachable)) abort (403);
             if (!Auth::user()->can('gd-bearbeiten')) abort(403);
             if (!Auth::user()->writableCities->contains($attachment->attachable->city)) abort(403);
         } else {
+            if (!Auth::user()->can('update', $attachment->attachable->service)) abort(403);
             if (!Auth::user()->can('gd-kasualien-bearbeiten')) abort(403);
             if (!Auth::user()->writableCities->contains($attachment->attachable->service->city)) abort(403);
         }
