+++ resources/views/funerals/edit.blade.php
@@ -53,8 +53,13 @@
 @endsection
 
 @section('scripts')
-    <script>var attachments = {{ count($funeral->attachments) }};</script>
-    <script src="{{ asset('js/pfarrplaner/attachments.js') }}"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
+    <script>
+        var commentRoute = '{{ route('comments.store') }}';
+        var commentOwner = '{{ $funeral->id }}';
+        var commentOwnerClass = 'App\\Funeral';
+    </script>
+    <script src="{{ asset('js/pfarrplaner/comments.js') }}"></script>
     <script>
         function setFieldStates() {
             if ($('#selType_input option:selected').val() == 'Urnenbeisetzung') {
