+++ resources/views/weddings/edit.blade.php
@@ -130,6 +130,13 @@
 @section('scripts')
     <script>var attachments = {{ count($wedding->attachments) }};</script>
     <script src="{{ asset('js/pfarrplaner/attachments.js') }}"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
+    <script>
+        var commentRoute = '{{ route('comments.store') }}';
+        var commentOwner = '{{ $wedding->id }}';
+        var commentOwnerClass = 'App\\Wedding';
+    </script>
+    <script src="{{ asset('js/pfarrplaner/comments.js') }}"></script>
     <script>
         $(document).ready(function () {
             $('#btnRemoveAttachment').click(function () {
